🚚 Rename dev to core
