+++ lamindb/_feature_set.py
@@ -7,7 +7,7 @@ from lnschema_core import Feature, FeatureSet, Registry, ids
 from lnschema_core.types import FieldAttr, ListLike
 
 from lamindb._utils import attach_func_to_class_method
-from lamindb.dev.hashing import hash_set
+from lamindb.core.hashing import hash_set
 
 from . import _TESTING
 from ._feature import convert_numpy_dtype_to_lamin_feature_type
