+++ lamindb/core/__init__.py
@@ -64,8 +64,8 @@ from lnschema_core.models import (
 
 from lamindb._query_manager import QueryManager
 from lamindb._query_set import QuerySet, RecordsList
-from lamindb.dev._feature_manager import FeatureManager
-from lamindb.dev._label_manager import LabelManager
+from lamindb.core._feature_manager import FeatureManager
+from lamindb.core._label_manager import LabelManager
 
 from . import _data, datasets, exceptions, fields, types
 from ._mapped_collection import MappedCollection
