+++ lamindb/core/_mapped_collection.py
