+++ tests/test_data.py
@@ -1,5 +1,5 @@
 import lamindb as ln
-from lamindb.dev._data import add_transform_to_kwargs
+from lamindb.core._data import add_transform_to_kwargs
 
 
 def test_add_transform_to_kwargs():
