+++ tests/test_storage.py
@@ -6,11 +6,11 @@ import numpy as np
 import pandas as pd
 import pytest
 import zarr
-from lamindb.dev.storage import delete_storage
-from lamindb.dev.storage._backed_access import backed_access
-from lamindb.dev.storage._zarr import read_adata_zarr, write_adata_zarr
-from lamindb.dev.storage.file import read_adata_h5ad
-from lamindb.dev.storage.object import infer_suffix, write_to_file
+from lamindb.core.storage import delete_storage
+from lamindb.core.storage._backed_access import backed_access
+from lamindb.core.storage._zarr import read_adata_zarr, write_adata_zarr
+from lamindb.core.storage.file import read_adata_h5ad
+from lamindb.core.storage.object import infer_suffix, write_to_file
 
 
 @pytest.fixture
